saving current page films to data-base
