+++ src/js/api/movies-api.js
@@ -1,5 +1,6 @@
 import imagesApiService from './apiServicePlagin';
 import imageCardTpl from '../../templates/card-markup.hbs';
+import dataBase from '../components/local-storage-db';
 
 const imagesContainer = document.querySelector('#js-gallery');
 
@@ -7,8 +8,8 @@ function onSearch() {
   imagesApiService
     .fetchArticles()
     .then(data => {
-      const currentPage = getFilmData(data.results);
-
+      const currentPage = dataBase.getFilmData(data.results);
+      dataBase.saveCurrentPage(currentPage);
       createGallery(currentPage);
     })
     .catch(onFetchError);
@@ -23,23 +24,3 @@ function onFetchError(message) {
 }
 
 onSearch();
-
-function getFilmData(data) {
-  let array = data.map(el => {
-    let film = {
-      id: String(el.id),
-      title: el.title,
-      original_title: el.original_title,
-      poster_path: `http://image.tmdb.org/t/p/w300${el.poster_path}`,
-      vote_average: el.vote_average,
-      vote_count: el.vote_count,
-      overview: el.overview,
-      popularity: el.popularity,
-      release_date: el.release_date.slice(0, 4),
-      genres: 'action, comedy',
-    };
-
-    return film;
-  });
-  return array;
-}
