+++ src/js/components/local-storage-db.js
@@ -1,0 +1,31 @@
+export default {
+  CURRENT_PAGE: 'currentPage',
+  QUEUE: 'queue',
+  WATCHED: 'watched',
+
+  saveCurrentPage(array) {
+    localStorage.setItem(this.CURRENT_PAGE, JSON.stringify(array));
+  },
+
+  getFilmData(data) {
+    let array = data.map(el => {
+      let film = {
+        id: String(el.id),
+        title: el.title,
+        original_title: el.original_title,
+        poster_path: `http://image.tmdb.org/t/p/w300${el.poster_path}`,
+        vote_average: el.vote_average,
+        vote_count: el.vote_count,
+        overview: el.overview,
+        popularity: el.popularity,
+        release_date: el.release_date.slice(0, 4),
+        genres: 'action, comedy',
+        watched: false,
+        queue: false,
+      };
+
+      return film;
+    });
+    return array;
+  },
+};
